backend/requirements.txt
# Lie Detection Backend

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
mediapipe==0.10.9
opencv-python==4.9.0.80
numpy==1.26.3
httpx==0.26.0
python-dotenv==1.0.0
